start extracting git operations into their own package
